requirements.txt
langchain
langgraph
chainlit
copilotkit
newsapi-python
smtplib
